requirements.txt
fastapi
uvicorn
requests
google-api-python-client
pydantic
pydantic-settings
python-dotenv
pandas
streamlit
matplotlibdir 